project/using-flask-restful/requirements.txt
Flask-JWT-Extended
Flask-RESTful
Flask-SQLAlchemy
passlib
python-dotenv
